A complete rewrite of this LWRP
The previous LWRP used HTTP/S to interrogate and modify
opentstack-identity data. As the API is somewhat of a moving target,
migrate the register provider to include everything it had previously,
combine that with the credentials provider, and rework all the
identity calls to use the Identity (keystone) CLI tools.

This logic follows the same logic that was found in the previous
provider. In fact, code was recycled where appropriate - much of the
same log messages and branch logic remain.

The two functions that power this methodology are:
identity_command - calls the identity client with the parameters
supplied and returns stdout upon success. Upon failure, raise a
RuntimeException with the message being stderr.
identity_uuid - wrapper around identity_command that given a type
of resource to interrogate and an index for the data returned, will
provide the UUID of the target or nil if not found. Upon error, this
function will bubble up an exception.

Bumping the openstack-common dependency to 0.2.1.

All tests pass.

Change-Id: I4230f047ee9c26d076ad14c48afac73f0e92c4ad